backend/requirements.txt
# Friday Research Console - Backend Dependencies

# Web framework
fastapi>=0.111.0
uvicorn[standard]>=0.30.0

# Database
psycopg2-binary>=2.9.9

# Data validation (included with FastAPI, pinned for consistency)
pydantic>=2.7.0

# Local .env support (dev convenience)
python-dotenv>=1.0.1

# CORS middleware (included with FastAPI)
# starlette  # auto-installed with fastapi
